+++ source/Internal/Framework/Config/DataObject/ShopConfigurationSetting.php
@@ -11,13 +11,9 @@ namespace OxidEsales\EshopCommunity\Internal\Framework\Config\DataObject;
 
 class ShopConfigurationSetting
 {
-    public const MODULE_CLASS_EXTENSIONS = 'aModuleExtensions';
     public const MODULE_CLASS_EXTENSIONS_CHAIN = 'aModules';
     public const MODULE_CONTROLLERS = 'aModuleControllers';
-    public const MODULE_VERSIONS = 'aModuleVersions';
-    public const MODULE_PATHS = 'aModulePaths';
     public const MODULE_SMARTY_PLUGIN_DIRECTORIES = 'moduleSmartyPluginDirectories';
-    public const MODULE_EVENTS = 'aModuleEvents';
     public const ACTIVE_MODULES = 'activeModules';
 
     /**